backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Environment & Configuration
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
chromadb==0.4.22

# AI/ML (Skip torch, transformers as already installed)
sentence-transformers==2.3.1
google-generativeai==0.3.2

# PDF Processing
pdfminer.six==20221105
pypdf2==3.0.1

# Data Processing & Export
pandas==2.1.4
openpyxl==3.1.2
numpy<2.0.0

# Utilities
python-dateutil==2.8.2
